+++ packages/riverpod_analyzer_utils_tests/test/generator_provider_declaration_test.dart
@@ -4,6 +4,72 @@ import 'package:test/test.dart';
 import 'analyser_test_utils.dart';
 
 void main() {
+  testSource('Decode dependencies with syntax errors', source: '''
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+const deps = <ProviderOrFamily>[];
+
+@Riverpod(dependencies: deps)
+int first(FirstRef ref) => 0;
+
+@Riverpod(dependencies: )
+int second(SecondRef ref) => 0;
+
+@Riverpod(dependencies: [gibberish])
+int forth(ForthRef ref) => 0;
+
+@Riverpod(dependencies: [if (true) forth])
+int fifth(FifthRef ref) => 0;
+
+@Riverpod(dependencies: [int])
+int sixth(SixthRef ref) => 0;
+''', (resolver) async {
+    final result = await resolver.resolveRiverpodAnalyssiResult(
+      ignoreErrors: true,
+    );
+
+    final errors =
+        result.resolvedRiverpodLibraryResults.expand((e) => e.errors).toList();
+
+    expect(errors, hasLength(6));
+
+    expect(
+      errors[0].message,
+      '@Riverpod(dependencies: <...>) only support list literals (using []).',
+    );
+    expect(errors[0].targetNode?.toSource(), 'deps');
+
+    expect(
+      errors[1].message,
+      '@Riverpod(dependencies: <...>) only support list literals (using []).',
+    );
+    expect(errors[1].targetNode?.toSource(), '');
+
+    expect(
+      errors[2].message,
+      '@Riverpod(dependencies: [...]) only supports elements annotated with @riverpod as values.',
+    );
+    expect(errors[2].targetNode?.toSource(), 'gibberish');
+
+    expect(
+      errors[3].message,
+      '@Riverpod(dependencies: [...]) does not support if/for/spread operators.',
+    );
+    expect(errors[3].targetNode?.toSource(), 'if (true) forth');
+
+    expect(
+      errors[4].message,
+      'Unsupported dependency. Only functions and classes annotated by @riverpod are supported.',
+    );
+    expect(errors[4].targetElement.toString(), 'int sixth(dynamic ref)');
+
+    expect(
+      errors[5].message,
+      'Failed to parse dependency Type (int*)',
+    );
+    expect(errors[5].targetElement?.toString(), 'int sixth(dynamic ref)');
+  });
+
   testSource('Decode name', runGenerator: true, source: r'''
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -192,24 +258,14 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has no dependency',
       );
       expect(
-        provider.value.annotation.dependenciesNode,
-        null,
-        reason: '${provider.key} has no dependency',
-      );
-      expect(
         provider.value.annotation.element.allTransitiveDependencies,
-        null,
-        reason: '${provider.key} has no dependency',
-      );
-      expect(
-        provider.value.annotation.dependenciesNode,
         null,
         reason: '${provider.key} has no dependency',
       );
     }
     for (final provider in empty.entries) {
       expect(
-        provider.value.annotation.dependencies,
+        provider.value.annotation.dependencies?.dependencies,
         isEmpty,
         reason: '${provider.key} has an empty list of dependencies',
       );
@@ -224,14 +280,14 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has an empty list of dependencies',
       );
       expect(
-        provider.value.annotation.dependenciesNode?.toSource(),
+        provider.value.annotation.dependencies?.node.toSource(),
         'dependencies: []',
         reason: '${provider.key} has an empty list of dependencies',
       );
     }
     for (final provider in providers.entries) {
       expect(
-        provider.value.annotation.dependencies,
+        provider.value.annotation.dependencies?.dependencies,
         hasLength(2),
         reason: '${provider.key} has two explicit dependencies',
       );
@@ -246,7 +302,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has two explicit dependencies',
       );
       expect(
-        provider.value.annotation.dependencies?[0],
+        provider.value.annotation.dependencies?.dependencies?[0],
         isA<RiverpodAnnotationDependency>()
             .having(
               (e) => e.provider,
@@ -257,7 +313,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has `empty` as first dependency',
       );
       expect(
-        provider.value.annotation.dependencies?[1],
+        provider.value.annotation.dependencies?.dependencies?[1],
         isA<RiverpodAnnotationDependency>()
             .having(
               (e) => e.provider,
@@ -285,7 +341,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
       );
 
       expect(
-        provider.value.annotation.dependenciesNode?.toSource(),
+        provider.value.annotation.dependencies?.node.toSource(),
         'dependencies: [empty, EmptyNotifier]',
         reason: '${provider.key} has two dependencies',
       );
@@ -293,7 +349,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
 
     for (final provider in nesteds.entries) {
       expect(
-        provider.value.annotation.dependencies,
+        provider.value.annotation.dependencies?.dependencies,
         hasLength(2),
         reason: '${provider.key} has two explicit dependencies',
       );
@@ -313,7 +369,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has two explicit dependencies',
       );
       expect(
-        provider.value.annotation.dependencies?[0],
+        provider.value.annotation.dependencies?.dependencies?[0],
         isA<RiverpodAnnotationDependency>()
             .having(
               (e) => e.provider,
@@ -324,7 +380,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
         reason: '${provider.key} has `providerDependency` as first dependency',
       );
       expect(
-        provider.value.annotation.dependencies?[1],
+        provider.value.annotation.dependencies?.dependencies?[1],
         isA<RiverpodAnnotationDependency>()
             .having(
               (e) => e.provider,
@@ -358,7 +414,7 @@ class NestedDependencyNotifier extends _$NestedDependencyNotifier {
       );
 
       expect(
-        provider.value.annotation.dependenciesNode?.toSource(),
+        provider.value.annotation.dependencies?.node.toSource(),
         'dependencies: [providerDependency, ProviderDependencyNotifier]',
         reason: '${provider.key} has two dependencies',
       );